LayoutTests/imported/w3c/web-platform-tests/svg/animations/seeking-events-6-expected.txt


FAIL Seeking backwards does not dispatch 'repeatEvent' assert_equals: Expected beginEvent event, but got repeatEvent event instead expected "beginEvent" but got "repeatEvent"
